Separate 1D config code into a config::d1 module
In preparation for adding 2D and 3D capabilities to rems, this
commit adds a d1 namespace to the config module. The name is odd,
but names cannot start with integers in Rust, and conventionally
module names are lowercase, so we write d1 instead of 1D.
